Add profiles for packaging module
As packaging requires third-party tooling (e.g. Sphinx) it seems a good idea to trigger the build of `WaarpPackaging` and `WaarpPackaging6` only if a specific profile is enabled.
This commit adds 3 profiles as the packaging module to build depends on the target JRE. JRE 8 and 11 require the same module but for sake of consistency 2 profiles (`packaging-jre8` and `packaging-jre11`) have been added.